Move snippets usage tests to a separate test_usage module
