+++ wagtail/snippets/tests/test_snippets.py
@@ -25,7 +25,7 @@ from wagtail.admin.admin_url_finder import AdminURLFinder
 from wagtail.admin.forms import WagtailAdminModelForm
 from wagtail.admin.panels import FieldPanel, ObjectList, get_edit_handler
 from wagtail.blocks.field_block import FieldBlockAdapter
-from wagtail.models import Locale, ModelLogEntry, Page, ReferenceIndex, Revision
+from wagtail.models import Locale, ModelLogEntry, Revision
 from wagtail.signals import published, unpublished
 from wagtail.snippets.action_menu import (
     ActionMenuItem,
@@ -59,7 +59,6 @@ from wagtail.test.testapp.models import (
     AdvertWithTabbedInterface,
     DraftStateCustomPrimaryKeyModel,
     DraftStateModel,
-    GenericSnippetPage,
     MultiPreviewModesModel,
     RevisableChildModel,
     RevisableModel,
@@ -3795,135 +3794,6 @@ class TestSnippetOrdering(TestCase):
         self.assertLess(
             SNIPPET_MODELS.index(AlphaSnippet), SNIPPET_MODELS.index(ZuluSnippet)
         )
-
-
-class TestUsageCount(TestCase):
-    fixtures = ["test.json"]
-
-    @classmethod
-    def setUpTestData(cls):
-        super().setUpTestData()
-        output = StringIO()
-        management.call_command("rebuild_references_index", stdout=output)
-
-    def test_snippet_usage_count(self):
-        advert = Advert.objects.get(pk=1)
-        self.assertEqual(advert.get_usage().count(), 2)
-
-
-class TestUsedBy(TestCase):
-    fixtures = ["test.json"]
-
-    @classmethod
-    def setUpTestData(cls):
-        super().setUpTestData()
-        output = StringIO()
-        management.call_command("rebuild_references_index", stdout=output)
-
-    def test_snippet_used_by(self):
-        advert = Advert.objects.get(pk=1)
-
-        self.assertIsInstance(advert.get_usage()[0], tuple)
-        self.assertIsInstance(advert.get_usage()[0][0], Page)
-        self.assertIsInstance(advert.get_usage()[0][1], list)
-        self.assertIsInstance(advert.get_usage()[0][1][0], ReferenceIndex)
-
-
-class TestSnippetUsageView(WagtailTestUtils, TestCase):
-    fixtures = ["test.json"]
-
-    def setUp(self):
-        self.user = self.login()
-
-    def test_use_latest_draft_as_title(self):
-        snippet = DraftStateModel.objects.create(text="Draft-enabled Foo, Published")
-        snippet.save_revision().publish()
-        snippet.text = "Draft-enabled Bar, In Draft"
-        snippet.save_revision()
-
-        response = self.client.get(
-            reverse(
-                "wagtailsnippets_tests_draftstatemodel:usage",
-                args=[quote(snippet.pk)],
-            )
-        )
-
-        # Should use the latest draft title in the header subtitle
-        self.assertContains(
-            response,
-            '<span class="w-header__subtitle">Draft-enabled Bar, In Draft</span>',
-        )
-
-    def test_usage(self):
-        # resave so that usage count gets updated
-        page = Page.objects.get(pk=2)
-        page.save()
-
-        gfk_page = GenericSnippetPage(
-            title="Foobar Title",
-            snippet_content_object=Advert.objects.get(pk=1),
-        )
-        page.add_child(instance=gfk_page)
-
-        response = self.client.get(
-            reverse(
-                "wagtailsnippets_tests_advert:usage",
-                args=["1"],
-            )
-        )
-        self.assertContains(response, "Welcome to the Wagtail test site!")
-        self.assertContains(response, "Foobar Title")
-        self.assertContains(response, "<td>Generic snippet page</td>", html=True)
-        self.assertContains(response, "Snippet content object")
-
-    def test_usage_without_edit_permission_on_snippet(self):
-        # Create a user with basic admin backend access
-        user = self.create_user(
-            username="basicadmin", email="basicadmin@example.com", password="password"
-        )
-        admin_permission = Permission.objects.get(
-            content_type__app_label="wagtailadmin", codename="access_admin"
-        )
-        user.user_permissions.add(admin_permission)
-        self.login(username="basicadmin", password="password")
-
-        response = self.client.get(
-            reverse(
-                "wagtailsnippets_tests_advert:usage",
-                args=["1"],
-            )
-        )
-        self.assertEqual(response.status_code, 302)
-
-    def test_usage_without_edit_permission_on_page(self):
-        # resave so that usage count gets updated
-        page = Page.objects.get(pk=2)
-        page.save()
-
-        # Create a user with edit access to snippets but not pages
-        user = self.create_user(
-            username="basicadmin", email="basicadmin@example.com", password="password"
-        )
-        admin_permission = Permission.objects.get(
-            content_type__app_label="wagtailadmin", codename="access_admin"
-        )
-        advert_permission = Permission.objects.get(
-            content_type__app_label="tests", codename="change_advert"
-        )
-        user.user_permissions.add(admin_permission)
-        user.user_permissions.add(advert_permission)
-        self.login(username="basicadmin", password="password")
-
-        response = self.client.get(
-            reverse(
-                "wagtailsnippets_tests_advert:usage",
-                args=["1"],
-            )
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertNotContains(response, "Welcome to the Wagtail test site!")
-        self.assertContains(response, "(Private page)")
-        self.assertContains(response, "<td>Page</td>", html=True)
 
 
 class TestSnippetHistory(WagtailTestUtils, TestCase):
